move delayed message fetcher to another file and add comments to code
